backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
ollama==0.3.3
langchain==0.3.1
langchain-community==0.3.1
sentence-transformers==3.1.1
faiss-cpu>=1.9.0
numpy==1.26.4
websockets==13.1
pydantic==2.9.2
python-multipart==0.0.9
aiofiles==23.2.1
